requirements.txt
attrs==24.2.0
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
constantly==23.10.4
coverage==7.6.9
cryptography==44.0.0
cssselect==1.2.0
defusedxml==0.7.1
filelock==3.16.1
greenlet==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.0
mock==5.1.0
packaging==24.2
parsel==1.9.1
Protego==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.0.0
pyOpenSSL==24.3.0
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.6.0
SQLAlchemy==2.0.36
tldextract==5.1.3
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.2.3
w3lib==2.2.1
zope.interface==7.2
